Created new file for helper functions and refactored main.py
